Rename utility modules and inventory plugin
The utility modules and inventory plugin still contained "cloudscale" in
their name. This is no longer needed as everything is in the
cloudscale_ch.cloud collection namespace.
