+++ plugins/modules/floating_ip.py
@@ -71,7 +71,7 @@ options:
       - Reverse PTR entry for this address.
       - You cannot set a reverse PTR entry for IPv6 floating networks. Reverse PTR entries are only allowed for single addresses.
     type: str
-extends_documentation_fragment: cloudscale_ch.cloud.cloudscale
+extends_documentation_fragment: cloudscale_ch.cloud.api_parameters
 '''
 
 EXAMPLES = '''
@@ -170,7 +170,7 @@ except ImportError:
     HAS_IPADDRESS = False
 
 from ansible.module_utils.basic import AnsibleModule, env_fallback, missing_required_lib
-from ..module_utils.cloudscale import (
+from ..module_utils.api import (
     AnsibleCloudscaleBase,
     cloudscale_argument_spec,
 )
