+++ plugins/modules/server_group.py
@@ -48,7 +48,7 @@ options:
     description:
       - Tags assosiated with the server groups. Set this to C({}) to clear any tags.
     type: dict
-extends_documentation_fragment: cloudscale_ch.cloud.cloudscale
+extends_documentation_fragment: cloudscale_ch.cloud.api_parameters
 '''
 
 EXAMPLES = '''
@@ -118,7 +118,7 @@ tags:
 '''
 
 from ansible.module_utils.basic import AnsibleModule
-from ..module_utils.cloudscale import (
+from ..module_utils.api import (
     AnsibleCloudscaleBase,
     cloudscale_argument_spec,
 )
