+++ plugins/modules/volume.py
@@ -68,7 +68,7 @@ options:
     description:
       - Tags associated with the volume. Set this to C({}) to clear any tags.
     type: dict
-extends_documentation_fragment: cloudscale_ch.cloud.cloudscale
+extends_documentation_fragment: cloudscale_ch.cloud.api_parameters
 '''
 
 EXAMPLES = '''
@@ -163,7 +163,7 @@ tags:
 '''
 
 from ansible.module_utils.basic import AnsibleModule
-from ..module_utils.cloudscale import (
+from ..module_utils.api import (
     AnsibleCloudscaleBase,
     cloudscale_argument_spec,
 )
